requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
SQLAlchemy==2.0.28
python-dotenv==1.0.1
Flask-WTF==1.2.2
pandas==2.2.1
openpyxl==3.1.2
xlrd>=2.0.1
chardet==5.2.0
gunicorn
psycopg2-binary
python-dateutil==2.9.0
requests>=2.31.0
cryptography>=42.0.0

# Testing
pytest==8.1.1
pytest-flask==1.3.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==24.3.0
responses==0.25.0
freezegun==1.4.0
pytest-html==4.1.1
